test: migrate zero-onboarding-setup to bdd
